requirements.txt
pymupdf==1.25.3
pymupdf4llm==0.0.17
markdownify==1.1.0
flask==3.1.0
gunicorn==23.0.0
google-cloud-storage==3.1.0
pytesseract==0.3.13
Pillow==11.1.0